[biabduction] Make test independent of Java specific concat
Summary:
This addresses a test difference between java versions. Infer's java tests are recorded with Java8 where string concat with a constant string uses `toString`. However, if tests are run on a machine where Java 11 is used, string concat is done via `makeConcatWithConstants` which causes tests to fail.

As a workaround, we replace the test so that Java version dependent string concat is not used.

Reviewed By: ngorogiannis

Differential Revision: D27394621

fbshipit-source-id: dfe1af2ac