[clang] add support for VarTemplateSpecializationDecl
Summary:
This was added in C++14. Was investigating how SIOF dealt with this but
it turns out it already does the right thing as the translation unit of
global variable templates shows up as the place they are instantiated
(not the one where they are declared), which works well for SIOF
checking.

Reviewed By: da319

Differential Revision: D27500998

fbshipit-source-id: b8b9b9c48