[sledge] Revise type of Domain join operation, it is not partial
Summary:
The only domain with a partial join is the lifting of a predicate
domain to a relation one, where the entry states are required to be
equal. This situation now indicates a programming error in the
analysis, rather than something that the domain should be responsible
for. Therefore this diff changes that check to an assertion and
simplifies the remaining join operations which are all total.

Reviewed By: jvillard

Differential Revision: D27828763

fbshipit-source-id: ec52ff741