[refactor] invert ProcAttributes -> Pvar dependency
Summary:
Just moving stuff around.

This is possibly useful for making Pvar depend on ProcAttributes for
other things, eg checking if a pvar is captured by a procedure (which
would be awkward to have in the API of ProcAttributes and not Pvar).
Overall it forced me to move a few other things around in a way that I
feel makes more sense anyway.

Reviewed By: skcho

Differential Revision: D28091497

fbshipit-source-id: 367a1f17c