[cost] Fix cost models that accept blocks to take captured vars into account
Summary:
Function calls that accept blocks as arguments may have additional arguments for the captured variables of the block. Cost models for these functions only considered the case where the block arguments didn't capture variables. This diff extends the model so that we handle captured variable case.

This fixes some FNs in the analysis.

Reviewed By: skcho

Differential Revision: D28183071

fbshipit-source-id: 6a045e80e