[pulse] correct order of calling context in trace
Summary:
The order was reversed when printing the trace, leading to confusion.

Also make sure we indicate which part of the trace we are printing when
there is more than one part (either context + access or invalidation +
access, or all three).

Also start nesting at <calling context length> to better represent the
role of the calling context visually.

Reviewed By: da319

Differential Revision: D28329263

fbshipit-source-id: b691fb1f4