[erl-frontend] Build names environment for resolving functions
Summary:
Collect imports and exports in a data structure ("names environment")
that is easy to look up.

Background:
A function call f(a1,...,an) is shorthand for m:f(a1,...,an) if there is
a -import(m, [..., f/n, ...]); otherwise it is shorthand for c:f(a1,...,an)
where c is the current module. There is an implicit import of the
special "erlang" module. Any ambiguity (e.g., imported twice, or
imported and local) is an error.  Also, if there is a -export([...,
f/n,...]) then f/n should be marked as public (ProcAttributes)

Reviewed By: jvillard

Differential Revision: D28290252

fbshipit-source-id: f6d777eb6