[merge] forked global tenv merge
Summary:
In Buck/Java the global type environments of each buck target captured need to be merged. So do the capture DBs. These two tasks can be done concurrently, as both have a computation and an I/O component, and interleaving them should improve perf.

Indeed, profiling the merge process with `offcputime.py` and `cpudist.py` (BPF tools) showed a significant amount of off-cpu time in tests (>40%) as well as a distribution of timings for off-cpu intervals that agrees with IO on a fast medium (ssd).

This diff forks a process to merge the type environments while doing the DB merge as normal. Initial results show an almost 2x improvement.

Reviewed By: skcho

Differential Revision: D28438808

fbshipit-source-id: 89c96f25b