[pulse] report leaks as soon as values become unreachable
Summary:
This makes reports more readable: they were all at the end of functions,
currently.

This is actually quite tricky to do as it involves detecting which
locations are unreachable.

Some of this logic can/should probably be shared with
`AbductiveDomain.discard_unreachable` but at the moment that's not the
case.

Reviewed By: skcho

Differential Revision: D28382590

fbshipit-source-id: bd4239a0c