[pulse] values equal to live values are not dead
Summary:
This fixes a memory leak false positive. When collecting unreachable
values we should be careful to take the equality relation into account.
Equal values are normally canonicalised but only with respect to "known"
equalities. This makes sure variables that are live thanks to the
"pruned" equalities are not discarded from the state.

Reviewed By: skcho

Differential Revision: D28382642

fbshipit-source-id: 2b898d754