[pulse] Revise semantics of TextUtils.is_empty
Summary:
There are two changes:

1/ Fix incorrect merge of two cases (empty and not-empty)

```
let<*> astate_equal_zero = ... in
let<*> astate_not_zero = ... in
[Ok (ContinueProgram astate_equal_zero); Ok (ContinueProgram astate_not_zero)]
```

This was an incorrect merge, because if there is an error in the former
case `astate_equal_zero`, it doesn't even evaluate the latter case
`astate_not_zero`.

2/ Cover all cases precisely. There are actually three cases to cover:

* `TextUtils.is_empty(null) = true`
* `TextUtils.is_empty("") = true`
* `TextUtils.is_empty("abc") = false`

However, in the previous model, it missed the second case.

Reviewed By: jvillard

Differential Revision: D29393579

fbshipit-source-id: e59d9a60d