[erl-frontend] First Erlang test with warnings
Summary:
The gist of this diff is a set of Pulse models for the built-ins
that have been put in Sil by the Erlang frontend:
  - __erlang_make_nil, for creating an empty list like []
  - __erlang_make_cons, for creating a nonempty list like [1|[]]
  - __erlang_pattern_fail, which marks that no pattern matched
The models for the first two update the abstract state. The model for
the third generates a reportable error.

The diff also includes a test.

And also a few bugfixes:
  - a type ErlangAny was used sometimes instead of ErlangAny*
  - a load that was meant to be n=13 was instead n=*13;
    changed to (*tmp=13; n=*tmp)
  - the Makefile's guard for the rebar build-system test was inside the large
    guard for java analyzers; of course, we don't need java to run rebar

Reviewed By: jvillard

Differential Revision: D29230161

fbshipit-source-id: c8fd6d88a