[pulse] better trace when realising latent invalid accesses
Summary:
"glue" together the trace from the caller with the trace from the callee
to form the access trace when applying LatentInvalidAccess summaries.
This is consistent with what happens in function calls and allows Pulse
to better keep track of the history of the value. The goal is to prevent
bugs from getting filtered out because the access trace doesn't contain
the invalidation event for the value.

Reviewed By: da319

Differential Revision: D29550320

fbshipit-source-id: b600c188d