[pulse] avoid some FPs due to incompleteness in the arihtmetic engine
Summary:
Before, atoms we didn't understand but could be a contradiction became
"true" if they were only about dead variables. This is unsound for
incorrectness logic, and introduces false positives.

Be more conservative and try to identify when that could happen and
prune these paths instead, thus regaining under-approximation.

This is a bit ad-hoc: we expect reasoning on equalities and
disequalities to be precise enough, but we know that inequalities have
very little solving power at the moment (eg we don't detect `x < 0, x >
0 => false`). In addition, we remark that having *one* `<` atom can
never be a contradiction. So, we claim a possible contradiction whenever
we find *two* (>1) atoms containing inequalities and involving the same
variable.

Reviewed By: skcho

Differential Revision: D29232544

fbshipit-source-id: ff91eb1e4