[pulse] Evaluate (e+0) as e
Summary: Pulse evaluated (e1+e2) as a new value that is a sum of evaluation results of e1 and e2.  However, when e2 is known as zero, we can have a simpler evaluation result without introducing the new value. This diff returns the evaluation result of e1 when e2 is known as zero.

Reviewed By: ngorogiannis

Differential Revision: D29736444

fbshipit-source-id: d5ce5a60e