[procdesc] look at capture DB only -- cost
Summary:
`AnalysisCallbacks.get_proc_desc` potentially introduces non-determinism because it first looks for the `proc_desc` in the capture DB and if absent it tries to retrieve it from a summary.

The problem is the two proc descs are generally different, as the one in the summary is preanalysed and the one in the capture DB is not.

Unless the caller specifically needs a preanalysed `procdesc`, calls to `AnalysisCallbacks.get_proc_desc` should be replaced with `Procdesc.load`.

Reviewed By: skcho

Differential Revision: D29826482

fbshipit-source-id: f8c7cb243