[ondemand] only look at capture DB for attributes
Summary:
Ondemand looks at the attributes in the summary of a function in preference to those in the capture DB, in order to decide whether a function needs to be analysed. This is unnecessary and introduces non-determinism (whether the summary has been written is a race condition; attributes may differ from a pre-analysed CFG vs captured).

Switch to only looking at the attributes column in the capture DB.

Reviewed By: da319

Differential Revision: D29932634

fbshipit-source-id: 8d7d1b145